chore: Update docker-compose configuration and refactor order cancellation interfaces
