+++ docker-compose.dev.yaml
@@ -9,3 +9,27 @@ services:
       - ./src:/app/src:ro
     environment:
       - NODE_ENV=development
+    depends_on:
+      - postgres-srv
+      - redis-srv
+    ports:
+      - "3000:3000"
+
+  postgres-srv:
+    image: postgres:15
+    environment:
+      POSTGRES_DB: food_delivery_dev
+      POSTGRES_USER: postgres
+      POSTGRES_PASSWORD: postgres
+    ports:
+      - "5433:5432"
+    volumes:
+      - postgres-db:/var/lib/postgresql/data
+
+  redis-srv:
+    image: redis:7-alpine
+    ports:
+      - "6379:6379"
+
+volumes:
+  postgres-db:
